test: migrate app import api controller tests to testcontainers
